perf: less requests to the slack server
This commit minimizes the time it takes to render a message. It uses the `@lru_cache` decorator for each function in `parse.py`. It also fetches the emoji(s) in one go and then substitutes each emoji. The previous logic was, in each `lambda`, fetch an emoji with a request to the Slack server.