refactor: partially move message-related functions
This commit partially moves message-related functions to the `MessagesManager` class itself. This helps match the separation of concerns paradigm, and further abstracts the handling of messages & channels. On an unrelated note, it also passes the entire channel to the `messages_updated` signal's handler function, instead of just the ID. This is to provide more information about the channel in the future.

Key changes:
- partially implemented seperation of concerns with message & channel handling
- the `messages_updated` signal's handler function now accepts the entire channel, not just it's ID