feat: show window when tray clicked or duplicate instance attempted
This commit adds functionality for showing the window when the tray icon is clicked, or detecting if an instance is already running and showing that when another instance is opened. To check if another instance is opened before the window is initialized (without `QSharedMemory`), a socket is opened, checking for a connection to the port of the Flask server which is run on instance openings. The Flask server utilizes signals to communicate with the PySide `MainWindow`, triggering a `window.show` call. It uses built-in signals on the Tray icon for the click & show window functionality.