feat: display names of who posted a message
This commit adds functionality that displays who posted a message. It also programmatically creates elements instead of using HTML strings, and uses custom mouse wheel and keyboard events to handle `Ctrl +/-` and `Ctrl + Mouse Wheel`. It uses the user ID to fetch the user details, such as their real name, caches it, and displays it.