perf: run IO operations in a separate thread
This commit enhances performance for profile picture loading, specifically the IO operations involved when caching. The previous logic was to simply write the binary data to file(s) and use that in the image displayed (if not already cached). Now, however, the IO operation runs in a thread and base64 is used when a file has not yet been created. This reduces the wait on IO operations, significantly enhancing performance.