feat: improve layout of a message
This commit improves the layout of message; specifically the profile picture. It makes the profile picture rounded using a mask which is a rounded rectangle, and applies it to the original message. This workaround is required, as the CSS property `border-radius` is not supported in a `QTextBrowser`. It also adds a newline after the image, so that the message is below the image.