fix: user info assignment in message processing
This commit fixes a bug where previous user's info was displayed for the next message. It does this by moving the `tasks` variable into the `for loop` scope. This is because when iterating over `tasks`, the user information is assigned to the fetched user information returned from the task. Thus, the last users` information is always displayed. It also makes a few changes, such as making `was_cached` True by default for easier checks, and making `users_pending_cache` a `dict` to access a user by their ID without looping over each user object and checking it's ID.