perf: use xxhash instead of md5
This commit uses xxhash, a faster, non-cryptographic hashing algorithm when caching. For some reason, this change highlighted (was there but never happened) an issue where the image was being changed to a string *while* caching, so the image was no longer bytes. It also fixes that. In addition, it now uses `_x{res}` instead of `image_{res}`