fix: cannot connect to the Flask server
This commit fixes an issue where you cannot connect to the Flask server. Essentially, the `httpretty` package used for request mocking during the development process must've been interfering with some sort of socket or similar low-level network libraries, and so the Flask server was unable to function correctly. It uses a wrapper class which decorates each `WebClient` non-internal request call with `httpretty` mocking when in development, and ends it afterwards, as to not break the Flask server.